+++ src/main/java/com/github/jlangch/venice/impl/util/reflect/ReflectionAccessor.java
@@ -43,6 +43,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import com.github.jlangch.venice.JavaMethodInvocationException;
+import com.github.jlangch.venice.SecurityException;
 import com.github.jlangch.venice.impl.util.Tuple2;
 import com.github.jlangch.venice.impl.util.Tuple4;
 import com.github.jlangch.venice.javainterop.ReturnValue;