+++ src/main/java/com/github/jlangch/venice/javainterop/IInterceptor.java
@@ -21,7 +21,9 @@
  */
 package com.github.jlangch.venice.javainterop;
 
+import com.github.jlangch.venice.SecurityException;
 import com.github.jlangch.venice.impl.util.MeterRegistry;
+
 
 /**
  * Defines a Venice interceptor