+++ src/main/java/com/github/jlangch/venice/javainterop/SandboxRecorder.java
@@ -27,6 +27,8 @@ import java.io.Writer;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
+import com.github.jlangch.venice.SecurityException;
+
 
 public class SandboxRecorder extends Interceptor {