+++ src/test/java/com/github/jlangch/venice/sandbox/Sandbox_VeniceFuture_Test.java
@@ -30,6 +30,7 @@ import java.io.File;
 
 import org.junit.jupiter.api.Test;
 
+import com.github.jlangch.venice.SecurityException;
 import com.github.jlangch.venice.Venice;
 import com.github.jlangch.venice.javainterop.Interceptor;
 import com.github.jlangch.venice.javainterop.SandboxInterceptor;